+++ CHANGELOG.md
@@ -20,8 +20,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 - Refactor requests for flattened collection table of contents to use function in the collection TOC service.
 - Refactor the download texts modal to get the current text title from the document head service.
 - Updated the development notes with brief descriptions of dependencies.
-- Deps: update Angular to 17.1.3.
-- Deps: update Ionic to 7.7.1.
+- Deps: update `@angular` to 17.1.3.
+- Deps: update `@ionic` to 7.7.1.
 - Deps: update `marked` to 12.0.0.
 - Deps: update `jasmine-core` to 5.1.2.
 - Deps: update `ng-extract-i18n-merge` to 2.10.0.
@@ -30,6 +30,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 ### Fixed
 
 - Correctly infer app mode on touch devices. Previously the app was set to mobile mode on all touchscreen devices, which was not optimal for e.g. laptops with touchscreens.
+
+### Removed
+
+- Deps: `xliff`.